requirements.txt
mysql-connector-python==8.0.32
pip==21.1.2
protobuf==3.20.3
setuptools==57.0.0
wheel==0.36.2